Improved set_scr_name function
Some code correction.

+++ sfall/Modules/Graphics.cpp
@@ -931,7 +931,7 @@ public:
 		0x4CB36B GNW95_SetPaletteEntries_
 	*/
 	HRESULT __stdcall SetEntries(DWORD a, DWORD b, DWORD c, LPPALETTEENTRY d) { // used to set palette for splash screen, fades, subtitles
-		if (!windowInit || c <= 0) return DDERR_INVALIDPARAMS;
+		if (!windowInit || (long)c <= 0) return DDERR_INVALIDPARAMS;
 		//palCounter++;
 
 		PALETTE* destPal = (PALETTE*)d;
@@ -939,7 +939,7 @@ public:
 		if (Graphics::GPUBlt) {
 			D3DLOCKED_RECT pal;
 			if (c != 256) {
-				RECT rect = { static_cast<long>(b), 0, static_cast<long>(b + c), 1 };
+				RECT rect = { (long)b, 0, (long)(b + c), 1 };
 				paletteTex->LockRect(0, &pal, &rect, D3DLOCK_NO_DIRTY_UPDATE);
 				paletteTex->AddDirtyRect(&rect);
 				b = 0;
